+++ package.json
@@ -105,6 +105,7 @@
     "@testing-library/jest-dom": "^6.4.2",
     "@testing-library/react": "^15.0.2",
     "@types/file-saver": "^2.0.5",
+    "@types/jest": "^29.5.14",
     "@types/lodash": "^4.14.194",
     "@types/node": "^20.14.14",
     "@types/plotly.js-dist-min": "^2.3.1",
